SwiftFormat - Step
## References:
[SwiftFormat](https://github.com/nicklockwood/SwiftFormat)

## Inspired by Prior Art:
https://github.com/kimdv/bitrise-step-swiftlint

## Goal:
Given SwiftFormat is a solution primarily targeted for code formatting, it lacks the enforcement protection for a repo. Although git hooks can make for autoformatting to occur automatically, it also is not a requirement to perform changes/PRs.

This step is an attempt to provide a strict format enforcement to ensure the code adhears to a format definition.